change: lock wallpaper final impl #122
